+++ changelog.md
@@ -10,7 +10,7 @@ All important changes to this plugin will be documented in this file.
 - SSO when opening the Multi Problem Selector for the Problem Pool
 
 ### Changed
-- Plugin now requires Moodle 4.3
+- Plugin now requires Moodle 4.3 or higher
 - Support Moodle 5.2
 - Multi Problem Selector now creates MUMIE Tasks via form submission instead of drag-and-drop
 - Settings configured on the form (points, deadline, container, etc.) are now applied to all tasks created via the Multi Problem Selector